DSND-2941: Add validation of request body and iTests
