+++ src/test/resources/requests/missing-required-fields-acsp-internal-request.json
@@ -1,0 +1,6 @@
+{
+  "acsp_full_profile": {
+  },
+  "internal_data": {
+  }
+}
